requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
chardet==4.0.0
click==8.1.6
colorzero==1.1
decorator==5.1.1
distro==1.5.0
font-fredoka-one==0.0.4
font-hanken-grotesk==0.0.2
font-intuitive==0.0.4
font-source-sans-pro==0.0.1
font-source-serif-pro==0.0.1
fonts==0.0.3
future==0.18.3
geocoder==1.38.1
gpiozero==1.6.2
html5lib==1.1
icalendar==4.0.3
idna==2.10
inky==1.5.0
lxml==4.6.3
numpy==1.19.5
picamera2==0.3.9
pidng==4.0.9
piexif==1.1.3
Pillow==8.1.2
python-apt==2.2.1
python-dateutil==2.8.1
python-prctl==1.7
pytz==2021.1
qrcode==6.1
ratelim==0.1.6
requests==2.25.1
RPi.GPIO==0.7.0
simplejpeg==1.6.4
six==1.16.0
slacker==0.14.0
smbus2==0.4.2
soupsieve==2.2.1
spidev==3.5
ssh-import-id==5.10
toml==0.10.1
Unidecode==1.3.6
urllib3==1.26.5
v4l2-python3==0.3.2
webencodings==0.5.1
wikiquotes==1.4
